完善了后端代码的注册，计算帖子数量，页数等功能

+++ src/main/java/com/bbs/service/impl/CommentServiceImpl.java
@@ -108,8 +108,11 @@ public class CommentServiceImpl implements ICommentService {
     }
 
     @Override
-    public void deleteByCommentId(Long commentId) {
+    public boolean deleteByCommentId(Long commentId) {
+        if(commentRepository.findByCommentId(commentId) == null)
+            return false;
         commentRepository.deleteByCommentId(commentId);
+        return commentRepository.findByCommentId(commentId) == null;
     }
 
     @Override
